update preventing touch + click events
I'm not sure this is the best method, but I've figured out a way to prevent combining touch and click events, through trial and error.

Since I had to add touchstart to get the dropdown to close with a whitespace touch, on mobile, I had to find a way to prevent combining the two events. Turning off click events, if a touchstart event is detected, seems to work, but it also means every click event has to be paired with a touch event.