create adjustedClientHeight
After adding content to the note, the scroll-up function stopped working. At certain screen widths, the scrollHeight value in the console was 1px higher than the combined scrollTop and clientHeight values, so the === comparison was failing. Not sure where the missing pixel goes, but I decided to add a 1px allowance to get the scroll button to work again.