note about h1 element
So, I initially decided against making this an h1, since it goes against best practice to hide an h1 element in a hidden popup. Especially since this heading isn't about the overarching project title. But, you can't really have an h2 without an h1, and the note was getting cluttered with paragraph tags, when more semantically appropriate tags exist.