clipboard onclick mobile test
This works in a different project, where the selected text is grabbed from a contenteditable div, with the .text() method. I'm assuming either the textarea or the required .val() method is the issue here.

This .toString() solution works on desktop. Checking to see if it finally works on mobile. If not, I'll replace the textarea with a div, like the other project.